Fix E0308: Remaining type mismatch errors
Fixed run_event and damage Some() wrapping in:
- perishsong, phantomforce, pluck, powder, saltcure

Fixed StatID conversion in photongeyser.rs
Fixed type casts in psywave.rs, stealthrock.rs
Fixed prioritize_action parameters in round.rs
Fixed is_weather parameter in shoreup.rs

Fixed 11+ E0308 instances across 10 files

Errors: 326 -> 324

Session progress: 357 -> 324 (33 errors fixed)

+++ src/data/move_callbacks/psywave.rs
@@ -24,7 +24,7 @@ pub fn damage_callback(battle: &mut Battle, pokemon_pos: (usize, usize), target_
         pokemon_pokemon.level
     };
 
-    let damage = (rand * level) / 100;
+    let damage = (rand * level as i32) / 100;
 
     EventResult::Number(damage)
 }
